FIX: Replace ConsumerStack by DeviceConsumerStack and ShapeConsumerStack in main_delete.sh FIX: Replace kubectl apply by delete in eks/delete_apps.sh Remove unused Redis index SHAPE_INDEX_TYPE Introduce lafleet_template.yaml and gitignore eksctl/lafleet.yaml Replace GET_CNT_FCN by REACT_CDN in handler_reactFrontend()
New script updateSourceCodeOnAws.sh
Add CODECOMMIT_DIR in set_env-vars.sh
downloadRepos.sh optimisation

+++ eks/delete_apps.sh
@@ -2,13 +2,18 @@
 
 . ./eks/setEnvVar.sh
 
-kubectl apply -f ./eks/devices-slow_deployment.yml
-kubectl apply -f ./eks/redisearch-query_service.yml
-kubectl apply -f ./eks/redisearch-performance-analytics-py_service.yml
-kubectl apply -f ./eks/sqsdeviceconsumer-toredisearch_deployment.yml
-kubectl apply -f ./eks/sqsshapeconsumer-toredisearch_deployment.yml
-kubectl apply -f ./eks/redisearch_service.yml
+kubectl get pods --all-namespaces
+if [ $? -eq 0 ]; then
+  kubectl delete -f ./eks/devices-slow_deployment.yml
+  kubectl delete -f ./eks/redisearch-query_service.yml
+  kubectl delete -f ./eks/redisearch-performance-analytics-py_service.yml
+  kubectl delete -f ./eks/sqsdeviceconsumer-toredisearch_deployment.yml
+  kubectl delete -f ./eks/sqsshapeconsumer-toredisearch_deployment.yml
+  kubectl delete -f ./eks/redisearch_service.yml
 
-kubectl apply -f https://github.com/kubernetes-sigs/metrics-server/releases/latest/download/components.yaml
+  kubectl delete -f https://github.com/kubernetes-sigs/metrics-server/releases/latest/download/components.yaml
 
-helm uninstall kubernetes-ingress --namespace haproxy-controller --wait
+  helm uninstall kubernetes-ingress --namespace haproxy-controller --wait
+else
+  echo "Using kubectl to delete apps on K8s failed, skipping"
+fi
